migrate tokens unit tests to pytest
